Added Keys to Kerbo and simple dash effect
Placeholder dash effect, can be replaces with sprites for different effects later on.

Made D key glow green when being used to display that it is being used.

The WASD keys are present and need to have their visibility toggled.

Will need to add an effect that covers Kerbo when he gets a new key!